Move tinyxml to header only. Cleaned up build system. Antioch now really header only.
