+++ src/parsing/include/antioch/read_reaction_set_data_xml.h
@@ -38,7 +38,7 @@
 #include "antioch/physical_constants.h"
 
 // XML
-#include "antioch/tinyxml2.h"
+#include "antioch/tinyxml2_imp.h"
 
 // C++
 #include <string>
